requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.1
Pillow>=10.3.0
gpiozero==2.0
RPLCD==1.3.0
smbus2==0.4.3
rpi-lgpio
